+++ BlogEngine/Models/Entities/Post.cs
@@ -2,8 +2,8 @@
 {
     public class Post : BaseEntity
     {
-        public int ID { get; set; }
-        public int CategoryID { get; set; }
+        public long ID { get; set; }
+        public long CategoryID { get; set; }
         public string Subject { get; set; }
         public string Body { get; set; }
     }
